mais refatorações, rota index adicionada, legibilidade do código otimizada
