+++ src/users.py
@@ -1,3 +1,14 @@
+from pydantic import BaseModel
+
+class User(BaseModel):
+    username: str
+    email:str | None = None
+    full_name: str | None = None
+    disabled: bool | None = None
+
+class UserInDB(User):
+    hashed_password: str
+
 fake_users_db = {
     "userteste": {
         "username": "userteste",
